moving partitioning configuration to a separate folder
