+++ samples/azure/azure-service-fabric-routing/Core_6/Shared/EndpointPartitioning/HardcodeReplyToAddressToLogicalAddress.cs
