+++ samples/azure/azure-service-fabric-routing/Core_6/Shared/Shared.csproj
@@ -51,10 +51,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="EndpointPartitioning\EndpointPartitioningExtensions.cs" />
     <Compile Include="Logger.cs" />
     <Compile Include="ReceiverSideDistribution\Configuration\PartitionAwareReceiverSideDistributionConfiguration.cs" />
     <Compile Include="ReceiverSideDistribution\PartitionMappingFailedException.cs" />
-    <Compile Include="ReceiverSideDistribution\Pipeline\HardcodeReplyToAddressToLogicalAddress.cs" />
+    <Compile Include="EndpointPartitioning\HardcodeReplyToAddressToLogicalAddress.cs" />
     <Compile Include="SenderSideDistribution\Configuration\PartitionAwareSenderSideDistributionConfiguration.cs" />
     <Compile Include="SenderSideDistribution\Configuration\SenderSideDistributionExtensions.cs" />
     <Compile Include="ReceiverSideDistribution\Pipeline\DistributeMessagesBasedOnHeader.cs" />
